% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{fixTroubledFile}
\alias{fixTroubledFile}
\title{Checks if the file is ready for lumi processing}
\usage{
fixTroubledFile(filename, path = "~/", sep = ";", dataset_name = "")
}
\arguments{
\item{filename}{Full filename of the data to check}

\item{path}{Output path for fixed dataset}

\item{sep}{Separator used for splitting data}

\item{dataset_name}{Name of the fixed dataset}
}
\value{
Filename of fixed or checked file
}
\description{
Check to see if the file contains numbers as row names. These files will
break the addControlData2lumi function call and must be fixed. We remove
all row names and return the filename of the new fixed file. If the file
is fine we don't do anything and return the original filename.
}
\examples{
control_filename <- fixTroubledFile(/path/to/control/dataset.csv)),
                                     path="/output/path",
                                     sep="\\t",
                                     dataset_name="control_dataset_name")

}
\seealso{
\code{\link{generateDatasetFile}}
}
\author{
Bjorn Fjukstad, \email{bjorn@cs.uit.no}
}
\keyword{fix}
\keyword{rownames}
