% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generateDatasetFile}
\alias{generateDatasetFile}
\title{Generate RData datasets from raw files.}
\usage{
generateDatasetFile(input_path, exprs_filename, overview_filename,
  samplesheet_filename, ctrl_data_filename, bad_samples = NULL,
  remove_suffix = NULL, output_path, output_name, single_file = FALSE)
}
\arguments{
\item{input_path}{Path to the folder, containing data}

\item{exprs_filename}{Gene expression filename}

\item{overview_filename}{Overview filename}

\item{samplesheet_filename}{Samplesheet filename}

\item{ctrl_data_filename}{Controls data filename}

\item{bad_samples}{Sample names that should be excluded}

\item{remove_suffix}{Remove samples with suffix in their name}

\item{output_path}{Path for the output result}

\item{output_name}{Name for output}

\item{single_file}{Save as separate files or all-in-one (single file)}
}
\description{
Creates RData files from expression file, control file, overview and
samplesheet.
}
\examples{
generateDatasetFile(input_path="/path/to/gene_expression_data_folder/",
                      exprs_filename="gene_expression_filename.csv",
                      overview_filename="overview_filename.csv",
                      samplesheet_filename="samplesheet_filename.csv"
                      ctrl_data_filename="controls_data_filename.csv"
                      remove_suffix=c("_a", "_b"),
                      bad_samples=c("Sample1", "Sample2"),
                      output_path="/path/to/folder/with/results/",
                      output_name="dataset_name",
                      single_file=TRUE)

}
\seealso{
\code{\link{makeLumi}}
}
\author{
Bjorn Fjukstad, \email{bjorn@cs.uit.no}
}
\keyword{dataset}
\keyword{expression}
\keyword{gene}
\keyword{generate}
