% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{getColumnFromData}
\alias{getColumnFromData}
\title{Subset column from data}
\usage{
getColumnFromData(dataset, column, levels = FALSE)
}
\arguments{
\item{column}{Column for the subset}

\item{levels}{Get levels (TRUE) or values (FALSE)}

\item{samples}{Dataset with samples}
}
\value{
subset of the data
}
\description{
Subset column from data
}
\examples{
getSamples
getColumnFromData(samples, "LPNR")

getDiseases
getColumnFromData(samples, "Disease", levels=TRUE)

getSubCohorts (Project)
getColumnFromData(samples, "Project", levels=TRUE)

}
\seealso{
\code{\link{selectDisease}, \link{selectDesign}, \link{selectSubCohort}}
}
\author{
Nikita Shvetsov, \email{nikita.shvetsov@uit.no}
}
\keyword{column}
\keyword{select}
\keyword{subset}
