% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getDesigns}
\alias{getDesigns}
\title{Set up available designs for NOWAC}
\usage{
getDesigns()
}
\value{
vector with available designs
}
\description{
Set up available designs for NOWAC
}
\seealso{
\code{\link{selectDesign}}
}
\author{
Bjorn Fjukstad, \email{bjorn@cs.uit.no}
}
\keyword{design}
