% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getHospital}
\alias{getHospital}
\title{Get hospital name}
\usage{
getHospital(code)
}
\arguments{
\item{code}{Integer code of the hospital}
}
\value{
string Hospital name
}
\description{
Helper function to translate the hospital code (a number between 1 and 11) to
a name, such as Tromso, Nodo or Molde.
}
\examples{
hospital_name <- getHospital(1)

}
\seealso{
\code{\link{nowaclite}}
}
\author{
Bjorn Fjukstad, \email{bjorn@cs.uit.no}
}
\keyword{hospital}
