% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{getOutroData}
\alias{getOutroData}
\title{Select outro data}
\usage{
getOutroData(samples, outro, sLNR = "LabNumber", oLNR = "labnr")
}
\arguments{
\item{samples}{Dataset with samples}

\item{outro}{Outer dataset (registry or questionaire)}

\item{sLNR}{Samples Id for samples dataset used for merging}

\item{oLNR}{Samples Id for outro dataset used for merging}
}
\value{
data related to evaluated samples
}
\description{
Select outro data (cancer registry or questionaire data)
}
\examples{
quest_data <- getOutroData(samples, two_page_quest)
registry_data <- getOutroData(samples, cancer_registry_2018, sLNR=LPNR, oLNR=LPNR)

}
\seealso{
\code{\link{getColumnFromData}}
}
\author{
Nikita Shvetsov, \email{nikita.shvetsov@uit.no}
}
\keyword{cancer}
\keyword{merging}
\keyword{outro}
\keyword{questionaires}
\keyword{registry}
