% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{makeLumi}
\alias{makeLumi}
\title{Creates lumi object from expression and control data file.}
\usage{
makeLumi(input_path, exprs_filename, ctrl_data_filename,
  bad_samples = NULL, remove_suffix = NULL, output_path = "~/")
}
\arguments{
\item{input_path}{Path to the folder, containing data}

\item{exprs_filename}{Gene expression filename}

\item{ctrl_data_filename}{Controls data filename}

\item{bad_samples}{Sample names that should be excluded}

\item{remove_suffix}{Remove samples with suffix in their name}

\item{output_path}{Path for the output result}
}
\value{
Lumi object (lobj)
}
\description{
Removes user-defined bad saples, remove suffixes
and returns the object. Gene expression and controls files should be
in folder, specified in \code{input_path} argument.
}
\examples{
lumi_object <- makeLumi("/path/to/gene_expression_data_folder/",
                         "gene_expression_filename.csv",
                         "controls_data_filename.csv",
                         bad_samples=c("Sample1","Sample2"),
                         remove_suffix=c("_a", "_b"),
                         output_path="/path/to/folder/with/results/")

}
\seealso{
\code{\link{generateDatasetFile}}
}
\author{
Bjorn Fjukstad, \email{bjorn@cs.uit.no}
}
\keyword{lumi}
\keyword{make}
\keyword{object}
