% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-nowac.R
\docType{package}
\name{nowaclite}
\alias{nowaclite}
\alias{nowaclite-package}
\title{nowaclite}
\description{
Lite version of NOWAC R package
}
\details{
This package documents several functions from the original package. It DOES NOT
contain any data, related to NOWAC project due to it's sensitivity.

To get access to data, preprocessing functions and pipeline please refer to
Nikita Shvetsov, \email{nikita.shvetsov@uit.no} or Bjorn Fjukstad, \email{bjorn@cs.uit.no}
}
