% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{selectDesign}
\alias{selectDesign}
\title{Select samples within specified design}
\usage{
selectDesign(samples, design)
}
\arguments{
\item{samples}{Dataset with samples}

\item{design}{String, identifiying design}
}
\value{
samples with specified design
}
\description{
Select samples within specified design
}
\examples{
samples <- selectDesign(samples, "case-control")

}
\seealso{
\code{\link{getColumnFromData}}
}
\author{
Bjørn Fjukstad, \email{bjorn@cs.uit.no}
}
\keyword{design}
\keyword{select}
