% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{selectDisease}
\alias{selectDisease}
\title{Select samples with specified diseases}
\usage{
selectDisease(samples, diseases)
}
\arguments{
\item{samples}{Dataset with samples}

\item{diseases}{Vector of diseases}
}
\value{
samples with specified diseases
}
\description{
Select samples with specified diseases
}
\examples{
diseases <- c("breast cancer")
samples <- selectDisease(samples, diseases)

}
\seealso{
\code{\link{getColumnFromData}}
}
\author{
Bjorn Fjukstad, \email{bjorn@cs.uit.no}
}
\keyword{diseases}
\keyword{select}
