% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{selectSubCohort}
\alias{selectSubCohort}
\title{Select samples within specified sub-cohorts}
\usage{
selectSubCohort(samples, sub_cohorts)
}
\arguments{
\item{samples}{Dataset with samples}

\item{design}{String, identifiying sub-cohorts}
}
\value{
samples within specified sub-cohorts
}
\description{
Select samples within specified sub-cohorts (also can be referred as Projects)
}
\examples{

samples <- selectSubCohort(samples, c("Hospital CC1", "Hospital CC2", "Hospital CC3"))

}
\seealso{
\code{\link{getColumnFromData}}
}
\author{
Bjorn Fjukstad, \email{bjorn@cs.uit.no}
}
\keyword{select}
\keyword{sub-cohort}
